requirements.txt
openai==1.55.1
python-dotenv
cairosvg
